requirements.txt
openai==1.61.1
pillow==11.0.0
python-dotenv==1.0.1